+++ schulcloud/adhoc_tasks/sodix/set_permissions.py
@@ -2,11 +2,9 @@
 
 import json
 import logging
-from typing import Literal, Dict, List
+from typing import List
 
-import requests
-import requests.auth
-
+import edusharing
 import util
 
 
@@ -31,52 +29,7 @@ class Blacklist:
             return self.all_groups
 
 
-class Node:
-
-    def __init__(self, obj: Dict):
-        try:
-            self.id: str = obj['ref']['id']
-            self.name: str = obj['name']
-            self.is_directory: bool = obj['isDirectory']
-        except KeyError:
-            raise RuntimeError(f'Could not parse node object: {obj}')
-        self.obj = obj
-
-    def __repr__(self):
-        return f'Node<{self.id}, {self.name}, is_dir={self.is_directory}>'
-
-
-class EdusharingAPI:
-
-    def __init__(self, base_url: str, username: str, password: str):
-        self.base_url = base_url + 'rest'
-        self.username = username
-        self.password = password
-        self.session = requests.Session()
-        self.session.auth = requests.auth.HTTPBasicAuth(self.username, self.password)
-
-    def make_request(self, method: Literal['GET', 'POST'], url: str, params: Dict[str, str]):
-        url = f'{self.base_url}{url}'
-        headers = {'Accept': 'application/json'}
-        return self.session.request(method, url, params=params, headers=headers)
-
-    def get_children(self, node_id: str) -> List[Node]:
-        url = f'/node/v1/nodes/-home-/{node_id}/children'
-        response = self.make_request('GET', url, {'maxItems': '500', 'skipCount': '0'})
-        if response.status_code == 200:
-            try:
-                return [Node(node) for node in response.json()['nodes']]
-            except KeyError:
-                raise RuntimeError(f'Could not parse response: {response.text}')
-        else:
-            raise RuntimeError(f'Unexpected response: {response.status_code}: {response.text}')
-
-    def set_permissions(self, node_id: str, groups: List[str], inheritance: bool = False):
-        # TODO: implement api call: https://edusharing.staging.hpi-schul-cloud.org/edu-sharing/swagger/#!/NODE_v1/setPermission
-        pass
-
-
-def find_node_by_name(api: EdusharingAPI, parent_id: str, child_name: str) -> Node:
+def find_node_by_name(api: edusharing.EdusharingAPI, parent_id: str, child_name: str) -> edusharing.Node:
     nodes = api.get_children(parent_id)
     for node in nodes:
         if node.name == child_name:
@@ -107,7 +60,10 @@ def main():
     environment = util.Environment(ENV_VARS, ask_for_missing=True)
     blacklist = create_blacklist_from_json('blacklist.json')
 
-    api = EdusharingAPI(environment['EDU_SHARING_BASE_URL'], environment['EDU_SHARING_USERNAME'], environment['EDU_SHARING_PASSWORD'])
+    api = edusharing.EdusharingAPI(
+        environment['EDU_SHARING_BASE_URL'],
+        environment['EDU_SHARING_USERNAME'],
+        environment['EDU_SHARING_PASSWORD'])
 
     sync = find_node_by_name(api, '-userhome-', 'SYNC_OBJ')
     sodix_spider = find_node_by_name(api, sync.id, 'sodix_spider')
